requirements.txt
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.15.1
chardet==3.0.4
click==8.0.4
cryptography==2.1.4
dataclasses==0.8
filetype==1.2.0
Flask==2.0.3
Flask-Cors==5.0.0
httplib2==0.9.2
idna==3.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauth==1.0.1
pkg-resources==0.0.0
pycparser==2.21
pycryptodome==3.21.0
requests==2.18.4
requests-unixsocket==0.1.5
simplejson==3.13.2
six==1.17.0
typing-extensions==4.1.1
urllib3==1.22
Werkzeug==2.0.3
zipp==3.6.0
